Init Go module, move main to cmd dir
To run, try `go run cmd/main.go apps.yml`
