+++ phenix/go.mod
@@ -1,0 +1,5 @@
+module phenix
+
+go 1.14
+
+require gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
